+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-# Unreleaed [xxxx-xx-xx]
+# `async-jsonrpc` crates
 
+## Version [Unreleased]
+  - `jsonrpc-types` 
+    - Update `async-tungstenite` to 0.14 ([#39])
+
+[#39]: https://github.com/koushiro/async-jsonrpc/pull/39